Split up DataRepository
