+++ gnd/src/main/java/com/google/android/gnd/repository/FeatureRepository.java
@@ -1,0 +1,121 @@
+/*
+ * Copyright 2018 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.gnd.repository;
+
+import com.google.android.gnd.model.Mutation;
+import com.google.android.gnd.model.Project;
+import com.google.android.gnd.model.feature.Feature;
+import com.google.android.gnd.model.feature.FeatureMutation;
+import com.google.android.gnd.persistence.local.LocalDataStore;
+import com.google.android.gnd.persistence.remote.RemoteDataEvent;
+import com.google.android.gnd.persistence.remote.RemoteDataStore;
+import com.google.android.gnd.persistence.sync.DataSyncWorkManager;
+import com.google.common.collect.ImmutableSet;
+import io.reactivex.Completable;
+import io.reactivex.Flowable;
+import io.reactivex.Maybe;
+import java8.util.Optional;
+import javax.inject.Inject;
+import javax.inject.Singleton;
+
+@Singleton
+public class FeatureRepository {
+  private final LocalDataStore localDataStore;
+  private final ProjectRepository projectRepository;
+  private final DataSyncWorkManager dataSyncWorkManager;
+
+  @Inject
+  public FeatureRepository(
+      LocalDataStore localDataStore,
+      RemoteDataStore remoteDataStore,
+      ProjectRepository projectRepository,
+      DataSyncWorkManager dataSyncWorkManager) {
+    this.localDataStore = localDataStore;
+    this.projectRepository = projectRepository;
+    this.dataSyncWorkManager = dataSyncWorkManager;
+
+    streamFeaturesToLocalDb(remoteDataStore);
+  }
+
+  /**
+   * Mirrors features in the current project from the remote db into the local db when the network
+   * is available. When invoked, will first attempt to resync all features from the remote db,
+   * subsequently syncing only remote changes.
+   */
+  private void streamFeaturesToLocalDb(RemoteDataStore remoteDataStore) {
+    // TODO: Move to Application or background service.
+    // TODO: Is this even working? If the returned Disposable is garbage collected this will be
+    // interrupted.
+    projectRepository
+        .getActiveProjectOnceAndStream()
+        .compose(Loadable::values)
+        .switchMap(remoteDataStore::loadFeaturesOnceAndStreamChanges)
+        .switchMap(event -> updateLocalFeature(event).toFlowable())
+        .subscribe();
+  }
+
+  private Completable updateLocalFeature(RemoteDataEvent<Feature> event) {
+    switch (event.getEventType()) {
+      case ENTITY_LOADED:
+      case ENTITY_MODIFIED:
+        return event.value().map(localDataStore::mergeFeature).orElse(Completable.complete());
+      case ENTITY_REMOVED:
+        // TODO: Delete features:
+        // localDataStore.removeFeature(event.getEntityId());
+        return Completable.complete();
+      case ERROR:
+        return Completable.error(event.error().get());
+      default:
+        return Completable.error(
+            new UnsupportedOperationException("Event type: " + event.getEventType()));
+    }
+  }
+
+  // TODO: Only return feature fields needed to render features on map.
+  // TODO(#127): Decouple from Project and accept id instead.
+  public Flowable<ImmutableSet<Feature>> getFeaturesOnceAndStream(Project project) {
+    return localDataStore.getFeaturesOnceAndStream(project);
+  }
+
+  // TODO(#127): Decouple Project from Feature and remove projectId.
+  // TODO: Replace with Single and treat missing feature as error.
+  public Maybe<Feature> getFeature(String projectId, String featureId) {
+    return projectRepository
+        .getActiveProjectOnceAndStream()
+        .compose(Loadable::values)
+        .firstElement()
+        .filter(project -> project.getId().equals(projectId))
+        .flatMap(project -> localDataStore.getFeature(project, featureId));
+  }
+
+  public Completable saveFeature(Feature feature) {
+    // TODO(#79): Assign owner and timestamps when creating new feature.
+    // TODO(#80): Update UI to provide FeatureMutations instead of Features here.
+    return localDataStore
+        .applyAndEnqueue(
+            FeatureMutation.builder()
+                .setType(Mutation.Type.CREATE)
+                .setProjectId(feature.getProject().getId())
+                .setFeatureId(feature.getId())
+                .setLayerId(feature.getLayer().getId())
+                .setNewLocation(Optional.of(feature.getPoint()))
+                // TODO(#101): Attach real credentials.
+                .setUserId("")
+                .build())
+        .andThen(dataSyncWorkManager.enqueueSyncWorker(feature.getId()));
+  }
+}
