+++ gnd/src/main/java/com/google/android/gnd/ui/home/HomeScreenViewModel.java
@@ -25,8 +25,9 @@ import com.google.android.gnd.model.Project;
 import com.google.android.gnd.model.feature.Feature;
 import com.google.android.gnd.model.feature.Point;
 import com.google.android.gnd.model.form.Form;
-import com.google.android.gnd.repository.DataRepository;
+import com.google.android.gnd.repository.FeatureRepository;
 import com.google.android.gnd.repository.Loadable;
+import com.google.android.gnd.repository.ProjectRepository;
 import com.google.android.gnd.ui.common.AbstractViewModel;
 import com.google.android.gnd.ui.common.Navigator;
 import com.google.android.gnd.ui.common.SharedViewModel;
@@ -42,7 +43,8 @@ import javax.inject.Inject;
 public class HomeScreenViewModel extends AbstractViewModel {
 
   private static final String TAG = HomeScreenViewModel.class.getSimpleName();
-  private final DataRepository dataRepository;
+  private final ProjectRepository projectRepository;
+  private final FeatureRepository featureRepository;
   private final Navigator navigator;
   /** The state and value of the currently active project (loading, loaded, etc.). */
   private final LiveData<Loadable<Project>> activeProject;
@@ -60,13 +62,17 @@ public class HomeScreenViewModel extends AbstractViewModel {
   @Nullable private Form selectedForm;
 
   @Inject
-  HomeScreenViewModel(DataRepository dataRepository, Navigator navigator) {
-    this.dataRepository = dataRepository;
+  HomeScreenViewModel(
+      ProjectRepository projectRepository,
+      FeatureRepository featureRepository,
+      Navigator navigator) {
+    this.projectRepository = projectRepository;
+    this.featureRepository = featureRepository;
     this.addFeatureDialogRequests = new SingleLiveEvent<>();
     this.openDrawerRequests = new SingleLiveEvent<>();
     this.featureSheetState = new MutableLiveData<>();
     this.activeProject =
-        LiveDataReactiveStreams.fromPublisher(dataRepository.getActiveProjectOnceAndStream());
+        LiveDataReactiveStreams.fromPublisher(projectRepository.getActiveProjectOnceAndStream());
     this.navigator = navigator;
     this.addFeatureClicks = PublishSubject.create();
 
@@ -74,7 +80,7 @@ public class HomeScreenViewModel extends AbstractViewModel {
         addFeatureClicks
             .switchMapSingle(
                 newFeature ->
-                    dataRepository
+                    featureRepository
                         .saveFeature(newFeature)
                         .toSingleDefault(newFeature)
                         .doOnError(this::onAddFeatureError)
@@ -84,7 +90,7 @@ public class HomeScreenViewModel extends AbstractViewModel {
   }
 
   public boolean shouldShowProjectSelectorOnStart() {
-    return dataRepository.getLastActiveProjectId().isEmpty();
+    return projectRepository.getLastActiveProjectId().isEmpty();
   }
 
   public MutableLiveData<Integer> getAddObservationButtonVisibility() {
@@ -160,7 +166,7 @@ public class HomeScreenViewModel extends AbstractViewModel {
 
   public void init() {
     // Last active project will be loaded once view subscribes to activeProject.
-    dataRepository.getLastActiveProjectId().ifPresent(dataRepository::activateProject);
+    projectRepository.getLastActiveProjectId().ifPresent(projectRepository::activateProject);
   }
 
   // TODO: Move to OfflineAreaViewModel
